Use new Go project layout to refactor the codes skip uploading /vendor files
