+++ configs/conf.go
@@ -1,0 +1,11 @@
+package configs
+
+var JwtKey = []byte("my_secret_key")
+
+var Host = "localhost:8080"
+var Version = "/api/v1"
+
+var DbDriver = "mysql"
+var DbUser = "root"
+var DbPass = "19900718qzyQZY"
+var DbName = "barley"
